FUNCTIONALITY: 1012650 (task) - not fully completed, but in good shape for a commit in case my computer decides to queer out on me again.

+++ ProcessObserver.vb
@@ -52,7 +52,7 @@ Namespace kCura.Windows.Process
 			WriteError(recordInfo, message)
     End Sub
 
-		Public Sub RaiseProgressEvent(ByVal totalRecords As Int32, ByVal totalRecordsProcessed As Int32, ByVal totalRecordsProccessedWithWarnings As Int32, ByVal totalRecordsProcessedWithErrors As Int32, ByVal startTime As DateTime, ByVal endTime As DateTime, Optional ByVal totalRecordsDisplay As String = Nothing, Optional ByVal totalRecordsProcessedDisplay As String = Nothing)
+		Public Sub RaiseProgressEvent(ByVal totalRecords As Int64, ByVal totalRecordsProcessed As Int64, ByVal totalRecordsProccessedWithWarnings As Int64, ByVal totalRecordsProcessedWithErrors As Int64, ByVal startTime As DateTime, ByVal endTime As DateTime, Optional ByVal totalRecordsDisplay As String = Nothing, Optional ByVal totalRecordsProcessedDisplay As String = Nothing)
 			RaiseEvent OnProcessProgressEvent(New ProcessProgressEvent(totalRecords, totalRecordsProcessed, totalRecordsProccessedWithWarnings, totalRecordsProcessedWithErrors, startTime, endTime, totalRecordsDisplay, totalRecordsProcessedDisplay))
 		End Sub
